refactor: consolidate Bot type and extract webhook handler
- Create centralized TownsBot type alias preserving command types
- Replace duplicate TownsBot interfaces with imported type
- Extract GitHub webhook handler to separate route file
- Improve code organization and type safety
